server/requirements.txt
fastapi[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
psycopg2
